// types/woocommerce.d.ts
export interface ProductCategory {
  id: number;
  name: string;
  slug: string;
}
// ...
export interface ProductCategoryFull extends ProductCategory {
  parent: number;
  description: string;
  display: CategoryDisplay;
  image: CategoryImage | null;
  menu_order: number;
  count: number;
  _links?: HATEOASLinks;
}

export type CategoryDisplay = 'default' | 'products' | 'subcategories' | 'both';

export interface CategoryImage {
  id: number;
  date_created: string;
  date_created_gmt: string;
  date_modified: string;
  date_modified_gmt: string;
  src: string;
  name: string;
  alt: string;
}
// ...
export interface HATEOASLinks {
  self: HATEOASLink[];
  collection: HATEOASLink[];
  [key: string]: HATEOASLink[];
}

export interface HATEOASLink {
  href: string;
}